shared/python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bpt-shared"
version = "1.0.0"
description = "Shared library for BPT microservices"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.10.0",
    "structlog>=23.2.0",
    "httpx>=0.25.0",
    "redis>=5.2.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-instrumentation-httpx>=0.49b0",
    "opentelemetry-instrumentation-redis>=0.49b0",
    "python-json-logger>=2.0.7",
    "pyjwt[crypto]>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[tool.ruff]
target-version = "py313"
line-length = 120
lint.select = ["E", "W", "F", "I", "N", "B", "C4", "UP"]
lint.ignore = ["E501", "B008"]

[tool.black]
line-length = 120
target-version = ['py313']

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]